orchestrator/requirements.txt
# Django и DRF
Django==4.2.9
djangorestframework==3.14.0
django-filter==23.5
drf-spectacular==0.26.5  # OpenAPI/Swagger documentation

# Celery и Redis
celery==5.3.4
redis==5.0.1

# HTTP и OData client
requests==2.31.0
urllib3==2.1.0
tenacity==8.2.3  # Для retry logic в OData client

# PostgreSQL
psycopg2-binary==2.9.9

# CORS для frontend
django-cors-headers==4.3.1

# Environment variables
django-environ==0.11.2
python-dotenv==1.0.0

# Encryption
cryptography==41.0.7

# Utilities
pytz==2023.3

# Development
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
ipython==8.17.2
